refactor: widget 디렉토리 정리
