+++ src/App.tsx
@@ -1,7 +1,6 @@
 import { BrowserRouter as Router } from "react-router-dom"
-import Header from "./widgets/ui/Header.tsx"
-import Footer from "./widgets/ui/Footer.tsx"
 import PostsManagerPage from "./pages/PostsManagerPage.tsx"
+import { Footer, Header } from "./widgets"
 
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
