--- src/widgets/filter/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./Filter"
