+++ src/widgets/index.ts
@@ -1,0 +1,3 @@
+export * from "./Filter"
+export * from "./Footer"
+export * from "./Header"
